Improve bat files.
git-svn-id: svn://svn.code.sf.net/p/zenlib/code/ZenLib/trunk@448 f098ebc7-0e84-4588-8ad0-e2271d839f10

+++ Release/Release_GNU_Prepare.bat
@@ -1,21 +1,21 @@
 @echo off
 
 rem --- Clean up ---
-del ZenLib_GNU_Prepare.7z
-rmdir ZenLib_GNU_Prepare /S /Q
+if exist ZenLib_GNU_Prepare.7z del ZenLib_GNU_Prepare.7z
+if exist ZenLib_GNU_Prepare    rmdir ZenLib_GNU_Prepare /S /Q
 mkdir ZenLib_GNU_Prepare
 
-@rem --- Copying : Sources ---
-xcopy ..\Source\ZenLib\*.h ZenLib_GNU_Prepare\Source\ZenLib\ /S
-xcopy ..\Source\ZenLib\*.c ZenLib_GNU_Prepare\Source\ZenLib\ /S
+rem --- Copying : Sources ---
+xcopy ..\Source\ZenLib\*.c   ZenLib_GNU_Prepare\Source\ZenLib\ /S
 xcopy ..\Source\ZenLib\*.cpp ZenLib_GNU_Prepare\Source\ZenLib\ /S
+xcopy ..\Source\ZenLib\*.h   ZenLib_GNU_Prepare\Source\ZenLib\ /S
 
-@rem --- Copying : Projects ---
+rem --- Copying : Projects ---
 xcopy ..\Project\GNU\* ZenLib_GNU_Prepare\Project\GNU\ /S
 
-@rem --- Copying : Release ---
-xcopy *.txt ZenLib_GNU_Prepare\Release\
+rem --- Copying : Release ---
 xcopy *.sh  ZenLib_GNU_Prepare\Release\
+xcopy *.txt ZenLib_GNU_Prepare\Release\
 
 rem --- Copying : Information files ---
 copy ..\*.txt ZenLib_GNU_Prepare\
